fix(ios): render attached images on engine prompt without reload
Mirror the CLI path's optimistic-insert pattern (sendPrompt) inside
submitEnginePrompt so the user's bubble — with inline image previews
when attachments are present — appears immediately on iOS instead of
only after the user leaves and re-enters the conversation. The desktop
renderer already does the equivalent optimistic insert in
engine-slice.ts; there is no engine_user_message wire event today
because prompt-pipeline-renderer.ts explicitly skips role='user' for
engine tabs. iOS now does the symmetric local insert.

The content string mirrors the desktop's marker format
(`[Attached <type>: <path>]`) so the existing engineUserBubble parser
(parseAttachmentSegments) renders each path via InlineAttachmentImage.
The attachment cache is already primed under the desktop path by the
upload-result handler, so previews render instantly with no extra
network round trip.

Engine Commands moved out of SessionViewModel+Commands.swift into a
new sibling file to keep both under the Swift 600-line cap once the
optimistic-insert comment block landed, per CLAUDE.md guidance: split
along natural seams rather than collapsing comments.
